added args to gomod removed args module and location
